lock is 32bits long. Using type "long" would promote 64bits on x86_64...
git-svn-id: https://svn.aros.org/svn/aros/trunk@53632 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/include/aros/types/spinlock_s.h
@@ -14,7 +14,7 @@ typedef struct {
             unsigned int        updating : 1;
         } slock;
         volatile unsigned char  block[4];
-        volatile unsigned long  lock;
+        volatile unsigned int   lock;
     };
 } __attribute__((__aligned__(128))) spinlock_t;
